+++ packages/amplify-category-auth/provider-utils/awscloudformation/triggers/function-template-dir/trigger-index.js
@@ -22,12 +22,12 @@ const modules = moduleNames.map(name => require(`./${name}`));
  * @type {import('@types/aws-lambda').APIGatewayProxyHandler}
  * 
  */
-exports.handler = async event => {
+exports.handler = async (event, context) => {
   /**
    * Instead of naively iterating over all handlers, run them concurrently with
    * `await Promise.all(...)`. This would otherwise just be determined by the
    * order of names in the `MODULES` var.
    */
-  await Promise.all(modules.map(module => module.handler(event)));
+  await Promise.all(modules.map(module => module.handler(event, context)));
   return event;
 };