+++ packages/amplify-category-function/src/__tests__/index.test.ts
@@ -1,6 +1,8 @@
 import { initEnv, isMockable } from '..';
 import sequential from 'promise-sequential';
-import { pathManager, stateManager } from 'amplify-cli-core';
+import { stateManager } from 'amplify-cli-core';
+import { getLocalFunctionSecretNames } from '../provider-utils/awscloudformation/secrets/functionSecretsStateManager';
+import { getAppId, secretsPathAmplifyAppIdKey } from '../provider-utils/awscloudformation/secrets/secretName';
 
 jest.mock('promise-sequential');
 jest.mock('amplify-cli-core', () => ({
@@ -16,11 +18,58 @@ jest.mock('amplify-cli-core', () => ({
   },
 }));
 
+jest.mock('../provider-utils/awscloudformation/secrets/functionSecretsStateManager');
+jest.mock('../provider-utils/awscloudformation/secrets/secretName');
+
+const getLocalFunctionSecretNames_mock = getLocalFunctionSecretNames as jest.MockedFunction<typeof getLocalFunctionSecretNames>;
+getLocalFunctionSecretNames_mock.mockReturnValue([]);
+const getAppId_mock = getAppId as jest.MockedFunction<typeof getAppId>;
+
 const sequential_mock = sequential as jest.MockedFunction<typeof sequential>;
 const stateManager_mock = stateManager as jest.Mocked<typeof stateManager>;
 
 describe('function category provider', () => {
+  beforeEach(() => {
+    jest.clearAllMocks();
+  });
+
   describe('initialize environment', () => {
+    it('sets secretsPathAmplfiyAppId in team-provider-info if function has secrets configured', async () => {
+      stateManager_mock.getTeamProviderInfo.mockReturnValueOnce({});
+      getLocalFunctionSecretNames_mock.mockReturnValueOnce(['TEST_SECRET']);
+      getAppId_mock.mockReturnValueOnce('testappid');
+      const contextStub = {
+        amplify: {
+          removeResourceParameters: jest.fn(),
+          getEnvInfo: jest.fn().mockReturnValue({ envName: 'dev' }),
+          getResourceStatus: () => ({
+            allResources: [
+              {
+                category: 'function',
+                resourceName: 'testFunction',
+              },
+            ],
+            resourcesToBeCreated: [],
+            resourcesToBeDeleted: [],
+            resourcesToBeUpdated: [],
+          }),
+        },
+      };
+      await initEnv(contextStub);
+      expect(stateManager_mock.setTeamProviderInfo).toBeCalledTimes(1);
+      expect(stateManager_mock.setTeamProviderInfo.mock.calls[0][1]).toMatchObject({
+        dev: {
+          categories: {
+            function: {
+              testFunction: {
+                [secretsPathAmplifyAppIdKey]: 'testappid',
+              },
+            },
+          },
+        },
+      });
+    });
+
     it('only initializes function category resources', async () => {
       const resourcesToBeCreated = [
         {
@@ -56,7 +105,7 @@ describe('function category provider', () => {
       const contextStub = {
         amplify: {
           removeResourceParameters: jest.fn(),
-          getEnvInfo: jest.fn(),
+          getEnvInfo: jest.fn().mockReturnValue({ envName: 'dev' }),
           getResourceStatus: () => ({
             allResources: [...resourcesToBeCreated, ...resourcesToBeDeleted, ...resourcesToBeUpdated],
             resourcesToBeCreated,
@@ -65,7 +114,6 @@ describe('function category provider', () => {
           }),
         },
       };
-      contextStub.amplify.getEnvInfo.mockImplementationOnce(() => 'dev');
       await initEnv(contextStub);
       expect(contextStub.amplify.removeResourceParameters.mock.calls.length).toBe(1);
       expect(contextStub.amplify.getEnvInfo.mock.calls.length).toBe(1);