+++ packages/amplify-e2e-core/src/categories/api.ts
@@ -1,7 +1,7 @@
 import * as fs from 'fs-extra';
 import _ from 'lodash';
 import * as path from 'path';
-import { addFeatureFlag, ExecutionContext, getCLIPath, nspawn as spawn, setTransformerVersionFlag, updateSchema } from '..';
+import { addFeatureFlag, checkIfBucketExists, ExecutionContext, getCLIPath, getProjectMeta, nspawn as spawn, setTransformerVersionFlag, updateSchema } from '..';
 import { multiSelect, singleSelect } from '../utils/selectors';
 import { selectRuntime, selectTemplate } from './lambda-function';
 import { modifiedApi } from './resources/modified-api-index';
@@ -835,3 +835,52 @@ export function modifyRestAPI(projectDir: string, apiName: string) {
   const indexFilePath = path.join(projectDir, 'amplify', 'backend', 'api', apiName, 'src', 'express', 'index.js');
   fs.writeFileSync(indexFilePath, modifiedApi);
 }
+
+export function cancelAmplifyMockApi(cwd: string, settings: any = {}): Promise<void> {
+  return new Promise<void>((resolve, reject) => {
+    spawn(getCLIPath(), ['mock', 'api'], { cwd, stripColors: true })
+      .wait('AppSync Mock endpoint is running')
+      .sendCtrlC()
+      .run((err: Error) => {
+        if (err && !/Killed the process as no output receive for/.test(err.message)) {
+          reject(err);
+        } else {
+          resolve();
+        }
+      });
+  });
+}
+
+export async function validateRestApiMeta(projRoot: string, meta?: any) {
+  meta = meta ?? getProjectMeta(projRoot);
+  expect(meta.providers.awscloudformation).toBeDefined();
+  const {
+    AuthRoleArn: authRoleArn,
+    UnauthRoleArn: unauthRoleArn,
+    DeploymentBucketName: bucketName,
+    Region: region,
+    StackId: stackId,
+  } = meta.providers.awscloudformation;
+  expect(authRoleArn).toBeDefined();
+  expect(unauthRoleArn).toBeDefined();
+  expect(region).toBeDefined();
+  expect(stackId).toBeDefined();
+  const bucketExists = await checkIfBucketExists(bucketName, region);
+  expect(bucketExists).toMatchObject({});
+
+  expect(meta.function).toBeDefined();
+  let seenAtLeastOneFunc = false;
+  for (let key of Object.keys(meta.function)) {
+    const { service, build, lastBuildTimeStamp, lastPackageTimeStamp, distZipFilename, lastPushTimeStamp, lastPushDirHash } =
+      meta.function[key];
+    expect(service).toBe('Lambda');
+    expect(build).toBeTruthy();
+    expect(lastBuildTimeStamp).toBeDefined();
+    expect(lastPackageTimeStamp).toBeDefined();
+    expect(distZipFilename).toBeDefined();
+    expect(lastPushTimeStamp).toBeDefined();
+    expect(lastPushDirHash).toBeDefined();
+    seenAtLeastOneFunc = true;
+  }
+  expect(seenAtLeastOneFunc).toBe(true);
+};