+++ packages/amplify-graphql-relational-transformer/src/schema.ts
@@ -355,10 +355,10 @@ function makeModelXFilterInputObject(
       const isList = isListType(field.type);
       let filterTypeName = baseType;
 
-      if (isScalar(field.type) || isList) {
-        filterTypeName = isList
-          ? ModelResourceIDs.ModelFilterListInputTypeName(baseType, true)
-          : ModelResourceIDs.ModelScalarFilterInputTypeName(baseType, false);
+      if (isScalar(field.type)) {
+        filterTypeName = ModelResourceIDs.ModelScalarFilterInputTypeName(baseType, false);
+      } else if (isList) {
+        filterTypeName = ModelResourceIDs.ModelFilterListInputTypeName(baseType, true);
       }
 
       return {