+++ packages/amplify-graphql-searchable-transformer/streaming-lambda/python_streaming_function.py
@@ -20,8 +20,8 @@ OPENSEARCH_REGION = os.environ['OPENSEARCH_REGION']
 DEBUG = True if os.environ['DEBUG'] == "1" else False
 OPENSEARCH_USE_EXTERNAL_VERSIONING = True if os.environ['OPENSEARCH_USE_EXTERNAL_VERSIONING'] == "true" else False
 
-# Multiple mapping types in an index is deprecated. Default to doc.
-DOC_TYPE = 'doc'
+# Multiple mapping types in an index is deprecated in OpenSearch ver 7.10+. Default to _doc.
+DOC_TYPE = '_doc'
 OPENSEARCH_MAX_RETRIES = 3 # Max number of retries for exponential backoff
 
 logger = logging.getLogger()