+++ packages/amplify-e2e-core/CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.6.12](https://github.com/aws-amplify/amplify-cli/compare/amplify-e2e-core@2.6.11...amplify-e2e-core@2.6.12) (2022-02-25)
+
+**Note:** Version bump only for package amplify-e2e-core
+
+
+
+
+
+## [2.6.11](https://github.com/aws-amplify/amplify-cli/compare/amplify-e2e-core@2.6.10...amplify-e2e-core@2.6.11) (2022-02-18)
+
+
+
+## 7.6.21 (2022-02-16)
+
+**Note:** Version bump only for package amplify-e2e-core
+
+
+
+
+
+## [2.6.10](https://github.com/aws-amplify/amplify-cli/compare/amplify-e2e-core@2.6.9...amplify-e2e-core@2.6.10) (2022-02-15)
+
+**Note:** Version bump only for package amplify-e2e-core
+
+
+
+
+
 ## [2.6.9](https://github.com/aws-amplify/amplify-cli/compare/amplify-e2e-core@2.6.5...amplify-e2e-core@2.6.9) (2022-02-10)